Allow multiple menus in extension.php

+++ src/Platform/Menus/Observer.php
@@ -34,97 +34,52 @@ class Observer {
 	 */
 	public function afterInstall(Extension $extension)
 	{
-		$children = $extension->menu;
+		if ( ! $menus = $this->extractMenus($extension)) return;
 
-		// If the attribute physically doesn't exist, we'll just
-		// skip. They may have chosen to create the items progamatically
-		// so we won't want to remove everything they've done.
-		if ( ! is_array($children))
+		foreach ($menus as $slug => $children)
 		{
-			return;
-		}
+			// Now, we'll prepare the children, which will fill in
+			// the blanks
+			$this->recursivelyPrepareChildren($extension, $children);
 
-		// If there's no children, we'll observe an uninstall
-		// event for the child. This'll remove any children
-		// in the database.
-		if (empty($children))
-		{
-			return $this->afterUninstall($extension);
-		}
-
-		// Now, we'll prepare the children, which will fill in
-		// the blanks
-		$this->recursivelyPrepareChildren($extension, $children);
-
-
-		// Now, we'll loop through our new array and we'll compare that
-		// to the children in the database. Any children in the database
-		// not matching our array which are assigned to this extension
-		// will be removed.
-		$slugs = array();
-		array_walk_recursive($children, function($value, $key) use (&$slugs)
-		{
-			if ($key == 'slug') $slugs[] = $value;
-		});
-
-		$existing = with(new Menu)
-		    ->newQuery()
-		    ->where('extension', '=', $extension->getSlug())
-		    ->whereNotIn('slug', $slugs)
-		    ->get();
-
-		foreach ($existing as $child)
-		{
-			$child->delete();
-		}
-
-		// Firstly, we'll purge the existing children (and any descendents)
-		// from the children array and put them in the existing array.
-		with($adminMenu = Menu::adminMenu())->findChildren();
-		$tree     = $adminMenu->toArray();
-		$existing = $tree['children'];
-		$this->recursivelyPurgeExisting($children, $existing);
-
-		$tree = array_merge($existing, $children);
-		$this->recursivelyStripAttributes($tree);
-
-		// Because we have just taken our existing hierarchy
-		// and added to it, we can save on the overhead of
-		// orphaning or deleting children as there'll never
-		// be any here. So, we'll just call this method as
-		// a speed improvement.
-		$adminMenu->mapTreeAndKeep($tree);
-	}
-
-	/**
-	 * Prepares children for the fancy mapping process that will occur,
-	 * including validating attributes.
-	 *
-	 * @param  Cartalyst\Extensions\Extension  $extension
-	 * @param  array  $children
-	 * @return void
-	 */
-	protected function recursivelyPrepareChildren(Extension $extension, &$children)
-	{
-		foreach ($children as &$child)
-		{
-			if ( ! isset($child['slug']))
+			// Now, we'll loop through our new array and we'll compare that
+			// to the children in the database. Any children in the database
+			// not matching our array which are assigned to this extension
+			// will be removed.
+			$slugs = array();
+			array_walk_recursive($children, function($value, $key) use (&$slugs)
 			{
-				throw new \InvalidArgumentException("All menu children require a slug to be mapped from extension.php, Extension [{$extension->getSlug()}] has one or more slugs missing from it's menu children.");
+				if ($key == 'slug') $slugs[] = $value;
+			});
+
+			$existing = with(new Menu)
+			    ->newQuery()
+			    ->where('extension', '=', $extension->getSlug())
+			    ->whereNotIn('slug', $slugs)
+			    ->get();
+
+			foreach ($existing as $child)
+			{
+				$child->delete();
 			}
 
-			$child['extension'] = $extension->getSlug();
+			// Firstly, we'll purge the existing children (and any descendents)
+			// from the children array and put them in the existing array.
+			$method = camel_case($slug).'Menu';
+			with($menu = Menu::$method())->findChildren();
+			$tree     = $menu->toArray();
+			$existing = $tree['children'];
+			$this->recursivelyPurgeExisting($children, $existing);
 
-			if ( ! isset($child['type'])) $child['type']             = 'static';
-			if ( ! isset($child['visibility'])) $child['visibility'] = 'always';
-			if ( ! isset($child['children'])) $child['children']     = array();
+			$tree = array_merge($existing, $children);
+			$this->recursivelyStripAttributes($tree);
 
-			if ( ! is_array($child['children']))
-			{
-				throw new \InvalidArgumentException("Menu child [{$child['slug']}] for Extension [{$extension->getSlug()}] has a children property that is not an array.");
-			}
-
-			$this->recursivelyPrepareChildren($extension, $child['children']);
+			// Because we have just taken our existing hierarchy
+			// and added to it, we can save on the overhead of
+			// orphaning or deleting children as there'll never
+			// be any here. So, we'll just call this method as
+			// a speed improvement.
+			$menu->mapTreeAndKeep($tree);
 		}
 	}
 
@@ -136,31 +91,36 @@ class Observer {
 	 */
 	public function afterUninstall(Extension $extension)
 	{
-		$children = $extension->menu;
+		if ( ! $menus = $this->extractMenus($extension)) return;
 
-		$slugs = array();
-
-		if (is_array($children) and ! empty($children))
+		foreach ($menus as $slug => $children)
 		{
-			array_walk_recursive($children, function($value, $key) use (&$slugs)
+			// Build up an array of all the slugs present in the children array
+			$slugs = array();
+			if (is_array($children) and ! empty($children))
 			{
-				if ($key == 'slug') $slugs[] = $value;
-			});
-		}
+				array_walk_recursive($children, function($value, $key) use (&$slugs)
+				{
+					if ($key == 'slug') $slugs[] = $value;
+				});
+			}
 
-		$query = with(new Menu)
-		    ->newQuery()
-		    ->where('extension', '=', $extension->getSlug());
+			$query = with(new Menu)
+			    ->newQuery()
+			    ->where('extension', '=', $extension->getSlug());
 
-		if (count($slugs) > 0)
-		{
-			$query->orWhereIn('slug', $slugs);
-		}
+			if (count($slugs) > 0)
+			{
+				$query->orWhereIn('slug', $slugs);
+			}
 
-		foreach ($query->get() as $child)
-		{
-			$child->refresh();
-			$child->delete();
+			// Refresh our nodes so they're not affected by deletions and
+			// remove them.
+			foreach ($query->get() as $child)
+			{
+				$child->refresh();
+				$child->delete();
+			}
 		}
 	}
 
@@ -201,6 +161,75 @@ class Observer {
 		{
 			$child->enabled = false;
 			$child->save();
+		}
+	}
+
+	/**
+	 * Extract all valid menus from an extension.php file,
+	 * returnin an array where the key is a menu slug and
+	 * the value is an array of children.
+	 *
+	 * @param  Cartalyst\Extensions\Extension  $extension
+	 * @return array
+	 */
+	protected function extractMenus(Extension $extension)
+	{
+		$menus = $extension->menus;
+
+		// If the attribute physically doesn't exist, we'll just
+		// skip. They may have chosen to create the items progamatically
+		// so we won't want to remove everything they've done.
+		if ( ! is_array($menus))
+		{
+			return;
+		}
+
+		$valid = array();
+
+		// If there's no children, we'll observe an uninstall
+		// event for the child. This'll remove any children
+		// in the database.
+		if (empty($menus))
+		{
+			foreach ($menus as $slug => $menu)
+			{
+				if ( ! is_array($menus)) continue;
+				$valid[$slug] = $menu;
+			}
+		}
+
+		return $valid;
+	}
+
+	/**
+	 * Prepares children for the fancy mapping process that will occur,
+	 * including validating attributes.
+	 *
+	 * @param  Cartalyst\Extensions\Extension  $extension
+	 * @param  array  $children
+	 * @return void
+	 */
+	protected function recursivelyPrepareChildren(Extension $extension, &$children)
+	{
+		foreach ($children as &$child)
+		{
+			if ( ! isset($child['slug']))
+			{
+				throw new \InvalidArgumentException("All menu children require a slug to be mapped from extension.php, Extension [{$extension->getSlug()}] has one or more slugs missing from it's menu children.");
+			}
+
+			$child['extension'] = $extension->getSlug();
+
+			if ( ! isset($child['type'])) $child['type']             = 'static';
+			if ( ! isset($child['visibility'])) $child['visibility'] = 'always';
+			if ( ! isset($child['children'])) $child['children']     = array();
+
+			if ( ! is_array($child['children']))
+			{
+				throw new \InvalidArgumentException("Menu child [{$child['slug']}] for Extension [{$extension->getSlug()}] has a children property that is not an array.");
+			}
+
+			$this->recursivelyPrepareChildren($extension, $child['children']);
 		}
 	}
